All operations running and displaying result on the screen

+++ app/src/main/java/com/example/calculator/Calculate.java
@@ -4,19 +4,19 @@ public class Calculate {
 
     private Calculate(){};
 
-    public static int add(int a, int b){
+    public static double add(double a, double b){
         return  a+b;
     }
 
-    public static float subtract(float a, float b){
+    public static double subtract(double a, double b){
         return a-b;
     }
 
-    public static float divide(float a, float b){
+    public static double divide(double a, double b){
         return a/b;
     }
 
-    public static float multiply(float a, float b){
+    public static double multiply(double a, double b){
         return a*b;
     }
 }
